Refactor the custom generators
